Extract composer prompt input handlers
