+++ src/app/components/workspace/composer/composer-prompt-input-panel.tsx
@@ -1,14 +1,8 @@
 import { Loader2 } from 'lucide-react'
-import type { ClipboardEvent, KeyboardEvent, RefObject } from 'react'
-import {
-  type ComposerSendMode,
-  eventToAcceleratorCandidates,
-  type KeybindingOverrides,
-  normalizeAccelerator,
-} from '../../../../../shared/keybindings'
+import type { ClipboardEvent, RefObject } from 'react'
+import type { ComposerSendMode, KeybindingOverrides } from '../../../../../shared/keybindings'
 import type { ComposerAttachment, DesktopActionInvoker } from '../../../desktop/types'
 import { getPathForFileQuery } from '../../../query/desktop-query'
-import { cn } from '../../../utils/cn'
 import type { SettingsOpenTarget } from '../../../views/settings/settingsTypes'
 import { ComposerDictationControls } from './composer-dictation-controls'
 import { ComposerFileMentionPanel } from './composer-file-mention-panel'
@@ -18,305 +12,12 @@ import {
   hasAttachmentHintInClipboardData,
 } from './composer-paste-attachments'
 import { ComposerSkillMentionPanel } from './composer-skill-mention-panel'
+import { SlashCommandPanel } from './composer-slash-command-panel'
 import { ComposerTextField } from './composer-text-field'
+import { handleComposerTextKeyDown } from './composer-text-keydown'
 import type { ComposerFileMentions } from './useComposerFileMentions'
 import type { ComposerSkillMentions } from './useComposerSkillMentions'
-import {
-  type ComposerSlashCommands,
-  getComposerSlashCommandGroupLabel,
-  getComposerSlashCommandOptionId,
-} from './useComposerSlashCommands'
-
-let graphemeSegmenter: Intl.Segmenter | null = null
-
-function getTextSegments(value: string) {
-  if (!('Segmenter' in Intl)) {
-    return Array.from(value, (segment, index) => ({ index, segment }))
-  }
-
-  graphemeSegmenter ??= new Intl.Segmenter(undefined, { granularity: 'grapheme' })
-  return [...graphemeSegmenter.segment(value)]
-}
-
-function SlashCommandOption({
-  command,
-  index,
-  previousCommand,
-  selected,
-  slashCommands,
-}: {
-  command: ComposerSlashCommands['commands'][number]
-  index: number
-  previousCommand: ComposerSlashCommands['commands'][number] | undefined
-  selected: boolean
-  slashCommands: ComposerSlashCommands
-}) {
-  const groupLabel = getComposerSlashCommandGroupLabel(command)
-  const previousGroupLabel = previousCommand
-    ? getComposerSlashCommandGroupLabel(previousCommand)
-    : null
-  return (
-    <div key={`${command.source}:${command.name}`}>
-      {previousGroupLabel === groupLabel ? null : (
-        <div className="px-2 pt-1.5 pb-1 text-[10px] font-semibold uppercase tracking-[0.14em] text-[color:var(--muted-2)]">
-          {groupLabel}
-        </div>
-      )}
-      <button
-        id={getComposerSlashCommandOptionId(index)}
-        type="button"
-        role="option"
-        aria-selected={selected}
-        className={cn(
-          'flex w-full items-center gap-2 rounded-lg px-2 py-1.5 text-left',
-          selected
-            ? 'bg-[color:var(--accent-bg)] text-[color:var(--text)]'
-            : 'text-[color:var(--muted)] hover:bg-[color:var(--surface-hover)] hover:text-[color:var(--text)]',
-        )}
-        onPointerEnter={() => slashCommands.setSelectedIndex(index)}
-        onMouseDown={(event) => event.preventDefault()}
-        onClick={() => slashCommands.selectCommand(command)}
-      >
-        <span className="shrink-0 font-mono text-[12px] text-[color:var(--text)]">
-          /{command.name}
-        </span>
-        {command.description ? (
-          <span className="min-w-0 truncate text-[12px]">{command.description}</span>
-        ) : null}
-      </button>
-    </div>
-  )
-}
-
-function SlashCommandPanel({
-  panelRef,
-  slashCommands,
-}: {
-  panelRef: RefObject<HTMLDivElement | null>
-  slashCommands: ComposerSlashCommands
-}) {
-  if (!slashCommands.open) return null
-  return (
-    <div
-      ref={panelRef}
-      id={slashCommands.listboxId}
-      role="listbox"
-      tabIndex={-1}
-      aria-label="Composer slash commands"
-      className="absolute right-0 bottom-full left-0 z-20 max-h-64 scroll-py-1.5 overflow-auto rounded-xl border border-[color:var(--border-strong)] bg-[color:var(--panel)] p-1.5 shadow-[var(--shadow)]"
-    >
-      {slashCommands.commands.length > 0 ? (
-        slashCommands.commands.map((command, index) => (
-          <SlashCommandOption
-            key={`${command.source}:${command.name}`}
-            command={command}
-            index={index}
-            previousCommand={slashCommands.commands[index - 1]}
-            selected={index === slashCommands.selectedIndex}
-            slashCommands={slashCommands}
-          />
-        ))
-      ) : (
-        <div className="px-2 py-2 text-[12px] text-[color:var(--muted)]">
-          {slashCommands.loading ? 'Loading commands…' : 'No matching commands'}
-        </div>
-      )}
-    </div>
-  )
-}
-
-type ComposerKeyDownInput = {
-  cancelDictation: () => Promise<void>
-  clearError: () => void
-  dictationActive: boolean
-  dictationTranscribing: boolean
-  inputLocked: boolean
-  onArrowNavigationOverride?: ((direction: 'previous' | 'next') => boolean) | undefined
-  onEscapeOverride?: (() => boolean) | undefined
-  onSubmitOverride?: (() => boolean) | undefined
-  slashCommands: ComposerSlashCommands
-  fileMentions: ComposerFileMentions
-  skillMentions: ComposerSkillMentions
-  setDraft: (value: string) => void
-  composerSendMode: ComposerSendMode
-  keybindings: KeybindingOverrides
-}
-
-function isCursorAtStart(textarea: HTMLTextAreaElement) {
-  return textarea.selectionStart === textarea.selectionEnd && textarea.selectionStart === 0
-}
-
-function isCursorAtEnd(textarea: HTMLTextAreaElement) {
-  return (
-    textarea.selectionStart === textarea.selectionEnd &&
-    textarea.selectionEnd === textarea.value.length
-  )
-}
-
-function handleOpenAutocompleteKeyDown(
-  event: KeyboardEvent<HTMLTextAreaElement>,
-  input: ComposerKeyDownInput,
-) {
-  return (
-    input.slashCommands.handleKeyDown(event) ||
-    input.fileMentions.handleKeyDown(event) ||
-    input.skillMentions.handleKeyDown(event)
-  )
-}
-
-function handleDeleteTextKey(
-  event: KeyboardEvent<HTMLTextAreaElement>,
-  setDraft: (value: string) => void,
-  clearError: () => void,
-) {
-  if (event.key !== 'Backspace' && event.key !== 'Delete') return false
-  if (event.altKey || event.ctrlKey || event.metaKey) return false
-  const textarea = event.currentTarget
-  const selectionStart = textarea.selectionStart
-  const selectionEnd = textarea.selectionEnd
-  const segments = getTextSegments(textarea.value)
-  const previousSegment = [...segments].reverse().find((segment) => segment.index < selectionStart)
-  const nextSegment = segments.find((segment) => segment.index > selectionEnd)
-  const deleteStart =
-    selectionStart === selectionEnd && event.key === 'Backspace'
-      ? (previousSegment?.index ?? 0)
-      : selectionStart
-  const deleteEnd =
-    selectionStart === selectionEnd && event.key === 'Delete'
-      ? (nextSegment?.index ?? textarea.value.length)
-      : selectionEnd
-  if (deleteStart === deleteEnd) return false
-  event.preventDefault()
-  const nextValue = `${textarea.value.slice(0, deleteStart)}${textarea.value.slice(deleteEnd)}`
-  setDraft(nextValue)
-  clearError()
-  window.requestAnimationFrame(() => textarea.setSelectionRange(deleteStart, deleteStart))
-  return true
-}
-
-function handleHorizontalBoundaryNavigation(
-  event: KeyboardEvent<HTMLTextAreaElement>,
-  onArrowNavigationOverride: ((direction: 'previous' | 'next') => boolean) | undefined,
-) {
-  if (
-    event.key === 'ArrowLeft' &&
-    isCursorAtStart(event.currentTarget) &&
-    onArrowNavigationOverride?.('previous')
-  ) {
-    event.preventDefault()
-    return true
-  }
-  if (
-    event.key === 'ArrowRight' &&
-    isCursorAtEnd(event.currentTarget) &&
-    onArrowNavigationOverride?.('next')
-  ) {
-    event.preventDefault()
-    return true
-  }
-  return false
-}
-
-function isComposerSubmitKey(event: KeyboardEvent<HTMLTextAreaElement>, mode: ComposerSendMode) {
-  if (event.key !== 'Enter') return false
-  if (mode === 'cmd-enter') return event.metaKey || event.ctrlKey
-  return !(event.shiftKey || event.metaKey || event.ctrlKey)
-}
-
-function matchesComposerCommandKey(
-  event: KeyboardEvent<HTMLTextAreaElement>,
-  input: ComposerKeyDownInput,
-  commandId: 'composer.newline' | 'composer.submit',
-) {
-  const override = input.keybindings[commandId]
-  if (override === null) return false
-  const accelerators =
-    typeof override === 'string'
-      ? [override]
-      : getComposerDefaultAccelerators(commandId, input.composerSendMode)
-  const candidates = new Set(eventToAcceleratorCandidates(event))
-  return accelerators.some((accelerator) => candidates.has(normalizeAccelerator(accelerator)))
-}
-
-function insertComposerNewline(
-  event: KeyboardEvent<HTMLTextAreaElement>,
-  setDraft: (value: string) => void,
-) {
-  const textarea = event.currentTarget
-  const selectionStart = textarea.selectionStart
-  const selectionEnd = textarea.selectionEnd
-  const nextCursor = selectionStart + 1
-  const nextValue = `${textarea.value.slice(0, selectionStart)}\n${textarea.value.slice(selectionEnd)}`
-  event.preventDefault()
-  setDraft(nextValue)
-  window.requestAnimationFrame(() => textarea.setSelectionRange(nextCursor, nextCursor))
-}
-
-function handleComposerNewlineCommand(
-  event: KeyboardEvent<HTMLTextAreaElement>,
-  input: ComposerKeyDownInput,
-) {
-  if (!matchesComposerCommandKey(event, input, 'composer.newline')) return false
-  if (event.key !== 'Enter' || event.altKey || event.ctrlKey || event.metaKey) {
-    insertComposerNewline(event, input.setDraft)
-  }
-  return true
-}
-
-function composerCommandHasOverride(
-  input: ComposerKeyDownInput,
-  commandId: 'composer.newline' | 'composer.submit',
-) {
-  return Object.hasOwn(input.keybindings, commandId)
-}
-
-function getComposerDefaultAccelerators(
-  commandId: 'composer.newline' | 'composer.submit',
-  mode: ComposerSendMode,
-) {
-  if (commandId === 'composer.submit') return mode === 'cmd-enter' ? ['CmdOrCtrl+Enter'] : ['Enter']
-  return mode === 'cmd-enter' ? ['Enter'] : ['Shift+Enter']
-}
-
-function handleComposerTextKeyDown(
-  event: KeyboardEvent<HTMLTextAreaElement>,
-  input: ComposerKeyDownInput,
-) {
-  if (input.inputLocked) {
-    event.preventDefault()
-    return
-  }
-  if (event.key === 'Escape' && (input.dictationActive || input.dictationTranscribing)) {
-    event.preventDefault()
-    void input.cancelDictation()
-    return
-  }
-  if (event.key === 'Escape' && input.onEscapeOverride?.()) {
-    event.preventDefault()
-    return
-  }
-  if (handleDeleteTextKey(event, input.setDraft, input.clearError)) {
-    return
-  }
-  if (handleOpenAutocompleteKeyDown(event, input)) {
-    return
-  }
-  if (handleComposerNewlineCommand(event, input)) {
-    return
-  }
-  if (
-    matchesComposerCommandKey(event, input, 'composer.submit') ||
-    (!composerCommandHasOverride(input, 'composer.submit') &&
-      isComposerSubmitKey(event, input.composerSendMode))
-  ) {
-    event.preventDefault()
-    if (!input.onSubmitOverride?.()) input.slashCommands.submit()
-    return
-  }
-  if (handleHorizontalBoundaryNavigation(event, input.onArrowNavigationOverride)) {
-    return
-  }
-}
+import type { ComposerSlashCommands } from './useComposerSlashCommands'
 
 type ComposerPromptInputPanelProps = {
   attachments: ComposerAttachment[]
